+++ src/android/android-wizard.ts
@@ -3,6 +3,7 @@ import * as fs from 'fs';
 // @ts-ignore - clack is ESM and TS complains about that. It works though
 import * as clack from '@clack/prompts';
 import * as path from 'path';
+import * as Sentry from '@sentry/node';
 import * as gradle from './gradle';
 import * as manifest from './manifest';
 import * as codetools from './code-tools';
@@ -69,6 +70,7 @@ async function runAndroidWizardWithTelemetry(
     clack.log.error(
       'No Gradle project found. Please run this command from the root of your project.',
     );
+    Sentry.captureException('No Gradle project found');
     await abort();
     return;
   }
@@ -77,10 +79,8 @@ async function runAndroidWizardWithTelemetry(
     gradle.selectAppFile(buildGradleFiles),
   );
 
-  const { selectedProject, authToken } = await getOrAskForProjectData(
-    options,
-    'android',
-  );
+  const { selectedProject, selfHosted, sentryUrl, authToken } =
+    await getOrAskForProjectData(options, 'android');
 
   // ======== STEP 1. Add Sentry Gradle Plugin to build.gradle(.kts) ============
   clack.log.step(
@@ -100,6 +100,7 @@ async function runAndroidWizardWithTelemetry(
       "Could not add Sentry Gradle plugin to your app's build.gradle file. You'll have to add it manually.\nPlease follow the instructions at https://docs.sentry.io/platforms/android/#install",
     );
   }
+  Sentry.setTag('gradle-plugin-added', pluginAdded);
 
   // ======== STEP 2. Configure Sentry SDK via AndroidManifest ============
   clack.log.step(
@@ -119,6 +120,7 @@ async function runAndroidWizardWithTelemetry(
       "Could not configure the Sentry SDK. You'll have to do it manually.\nPlease follow the instructions at https://docs.sentry.io/platforms/android/#configure",
     );
   }
+  Sentry.setTag('android-manifest-updated', manifestUpdated);
 
   // ======== STEP 3. Patch Main Activity with a test error snippet ============
   clack.log.step(
@@ -133,10 +135,13 @@ async function runAndroidWizardWithTelemetry(
     packageName = gradle.getNamespace(appFile);
   }
   const activityName = mainActivity.activityName;
+  Sentry.setTag('has-activity-name', !!activityName);
+  Sentry.setTag('has-package-name', !!packageName);
   if (!activityName || !packageName) {
     clack.log.warn(
       "Could not find Activity with intent action MAIN. You'll have to manually verify the setup.\nPlease follow the instructions at https://docs.sentry.io/platforms/android/#verify",
     );
+    Sentry.captureException('Could not find Main Activity');
   } else {
     const packageNameStable = packageName;
     const activityFile = traceStep('Find Main Activity Source File', () =>
@@ -151,6 +156,7 @@ async function runAndroidWizardWithTelemetry(
         "Could not patch main activity. You'll have to manually verify the setup.\nPlease follow the instructions at https://docs.sentry.io/platforms/android/#verify",
       );
     }
+    Sentry.setTag('main-activity-patched', activityPatched);
   }
 
   // ======== STEP 4. Add sentry-cli config file ============
@@ -164,11 +170,16 @@ async function runAndroidWizardWithTelemetry(
   );
 
   // ======== OUTRO ========
+  const issuesPageLink = selfHosted
+    ? `${sentryUrl}organizations/${selectedProject.organization.slug}/issues/?project=${selectedProject.id}`
+    : `https://${selectedProject.organization.slug}.sentry.io/issues/?project=${selectedProject.id}`;
+
   clack.outro(`
 ${chalk.greenBright('Successfully installed the Sentry Android SDK!')}
 
 ${chalk.cyan(
-  'You can validate your setup by launching your application and checking Sentry issues page afterwards',
+  `You can validate your setup by launching your application and checking Sentry issues page afterwards
+${issuesPageLink}`,
 )}
 
 Check out the SDK documentation for further configuration: