+++ src/nextjs/nextjs-wizard.ts
@@ -118,6 +118,8 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
   const sentryWebpackOptionsTemplate = getNextjsWebpackPluginOptionsTemplate(
     selectedProject.organization.slug,
     selectedProject.slug,
+    selfHosted,
+    sentryUrl,
   );
   const sentryBuildOptionsTemplate = getNextjsSentryBuildOptionsTemplate();
 
@@ -323,6 +325,13 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
       useClient: true,
     });
 
+    fs.mkdirSync(
+      path.join(process.cwd(), ...appLocation, 'sentry-example-page'),
+      {
+        recursive: true,
+      },
+    );
+
     await fs.promises.writeFile(
       path.join(
         process.cwd(),
@@ -340,9 +349,12 @@ export async function runNextjsWizard(options: WizardOptions): Promise<void> {
       )}.`,
     );
 
-    fs.mkdirSync(path.join(process.cwd(), ...appLocation, 'api'), {
-      recursive: true,
-    });
+    fs.mkdirSync(
+      path.join(process.cwd(), ...appLocation, 'api', 'sentry-example-api'),
+      {
+        recursive: true,
+      },
+    );
 
     await fs.promises.writeFile(
       path.join(