+++ src/nextjs/templates.ts
@@ -1,6 +1,8 @@
 export function getNextjsWebpackPluginOptionsTemplate(
   orgSlug: string,
   projectSlug: string,
+  selfHosted: boolean,
+  url: string,
 ): string {
   return `{
     // For all available options, see:
@@ -8,9 +10,8 @@ export function getNextjsWebpackPluginOptionsTemplate(
 
     // Suppresses source map uploading logs during build
     silent: true,
-
     org: "${orgSlug}",
-    project: "${projectSlug}",
+    project: "${projectSlug}",${selfHosted ? `\n    url: "${url}"` : ''}
   }`;
 }